Enhance app functionality and user experience
Implemented sound playback, improved weather data retrieval, and refined UI layout and styling. Added new resources and updated several files to support these features, ensuring a more interactive and user-friendly application.

+++ WeatherSample.cs
@@ -202,7 +202,7 @@ namespace sysinfo
 
 public class WeatherExtraInfromation
 {
-  public int UvIndex;
+  public double UvIndex;
   public string? WeatherInformation;
